pyproject.toml
[project]
name = "advent-of-code-libraries"
description = "All my libraries for Advent of Code"

requires-python = ">=3.12"
dynamic = ["version"]
dependencies = [
    "advent-of-code-data",
    "aocd-example-parser",
    "aocd==0.1",
    "gitpython==3.1.40",
    "ipdb==0.13.13",
    "loguru==0.7.2",
    "networkx==3.2.1",
    "numpy==2.1",
    "pyfunctional==1.4.3",
    "pytest==7.4.3",
    "regex==2023.12.25",
    "rich-argparse==1.4.0",
    "ruff>=0.8.1",
    "shapely==2.0.6",
    "sympy==1.12",
    "peek-python>=1.6.2",
    "igraph>=0.11.8",
]

[tool.setuptools.packages.find]
where = ["."]
#include = ["libs.*"]
#include = ["input"]

[project.scripts]
aocl = "aocl:cli"

[project.entry-points."adventofcode.user"]
"alesieur" = "aocl:mysolve"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = [
  "I",  # isort
]

[tool.ruff.lint.isort]
force-single-line= true

[tool.uv]
package = true
